added auth related functionality to auth.js
