+++ app/controllers/user.js
@@ -14,24 +14,6 @@ module.exports = {
             res.status(400).send({ error: error })
         }
     },
-    authenticateUser: async function (req, res, next) {
-        const email = req.body.email
-        const password = req.body.password
-        try {
-            const user = await User.findByCredentials(email, password)
-            const token = await user.generateAuthToken()
-            res.send({ user: user, token: token })
-        } catch (error) {
-            console.log('error ', error)
-            res.status(400).send({ error: error })
-        }
-    },
-    logout: function (req, res, next) {
-        res.json({ success: 'ok' })
-    },
-    logoutAll: function (req, res, next) {
-        res.json({ success: 'ok' })
-    },
     userProfile: function (req, res, next) {
         res.json({ success: 'ok' })
     },
